+++ addon/components/uic-context-menu.js
@@ -2,43 +2,11 @@
   @module ember-ui-components
 */
 import Ember from 'ember';
+import openContextMenu from 'ember-ui-components/helpers/open-context-menu';
 import layout from '../templates/components/uic-context-menu';
-import { getDimensions, getMousePosition, calculatePosition } from 'ember-ui-components/lib/fn';
 
-let _event;
-
-/*
-  ## setPosition
-
-  Set the left/top css properties of an element.
-
-  `element` should be a reference to an HTML element.  Either a string selector
-  that can be used with jQuery, or a jQuery selection object.
-
-  If `position` is not specified, then the current mouse position will be used.
-
-  `position` should be an Ember.Object with `x` and `y` properties.
-  Both `x` and `y` should be numbers
-
-  method setPosition
-  param {String|Object} element
-  param {Object} position
-*/
-function setPosition(element, position) {
-  let margin = 5;
-  if (typeof element === 'string') {
-    element = Ember.$(element);
-  }
-  if (!position) {
-    position = getMousePosition(window.event || _event);
-  }
-  let scrollBarWidth = (window.innerWidth - $(window).width());
-  let el = getDimensions(element);
-  element.css({
-    'left': calculatePosition( position.get('x') + 2, el.width, window.innerWidth - ( margin + scrollBarWidth)),
-    'top': calculatePosition( position.get('y'), el.height, window.innerHeight - margin)
-  });
-}
+const { computed } = Ember;
+const { alias } = computed;
 
 /**
   @class ContextMenuComponent
@@ -47,6 +15,8 @@ function setPosition(element, position) {
 export default Ember.Component.extend({
 
   layout,
+
+  contextMenuService: Ember.inject.service('context-menu'),
 
   /**
     @property classNames
@@ -92,52 +62,24 @@ export default Ember.Component.extend({
   /**
     @property showContextMenu
     @type {Boolean}
-    @default false
   */
-  showContextMenu: false,
+  showContextMenu: computed('contextMenuService.menu', 'elementId', function () {
+    return this.get('contextMenuService.menu') === this.get('elementId');
+  }),
 
   /**
     @property contextMenuParams
     @type {Object}
     @private
   */
-  contextMenuParams: null,
-
-  /**
-    @method init
-    @private
-  */
-  init() {
-    this._super(...arguments);
-    this.initContextMenuParams();
-  },
-
-  /**
-    @method initContextMenuParams
-    @private
-  */
-  initContextMenuParams() {
-    this.set('contextMenuParams', Ember.Object.create({
-      event: null
-    }));
-  },
-
-  /**
-    This method is passed to the `didInsertElement` hook of the `{{uic-content-mask}}`
-    component which wraps the `.uic-context-menu-container` element.
-    @method didInsertContextMenu
-  */
-  didInsertContextMenu() {
-    let element = this.$('.uic-context-menu-container');
-    setPosition(element);
-  },
+  contextMenuParams: alias('contextMenuService.contextMenuParams'),
 
   /**
     @method _closeContextMenu
     @private
   */
   _closeContextMenu() {
-    this.set('showContextMenu', false);
+    this.get('contextMenuService').close();
   },
 
   /**
@@ -166,10 +108,7 @@ export default Ember.Component.extend({
   */
   contextMenu(event) {
     event.preventDefault();
-    _event = event;
-    this.set('contextMenuParams.event', event);
-    this.set('showContextMenu', true);
-    this.$().focus(); // set focus so that keyUp/Down events can be recieved.
+    this.get('contextMenuService').open(this.get('elementId'), event);
     return false;
   },