+++ app/services/context-menu.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-ui-components/services/context-menu';