+++ tests/dummy/app/controllers/context-menu-component.js
@@ -2,6 +2,8 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
 
+  contextMenuService: Ember.inject.service('context-menu'),
+  
   actions: {
     newItem() {
       // Ember.Logger.log('newItem');
@@ -11,6 +13,16 @@ export default Ember.Controller.extend({
     },
     delete() {
       // Ember.Logger.log('delete');
-    }
+    },
+    doSomething() {
+      Ember.Logger.log('doSomething');
+    },
+    doSomethingElse() {
+      Ember.Logger.log('doSomethingElse');
+    },
+
+    // contextmenu(id) {
+    //   Ember.Logger.log('contextmenu', id);
+    // }
   }
 });