+++ tests/unit/helpers/open-context-menu-test.js
@@ -1,0 +1,8 @@
+import openContextMenu from 'dummy/helpers/open-context-menu';
+import { module, test } from 'qunit';
+
+module('Unit | Helper | open context menu');
+
+test('it works', function(assert) {
+  assert.equal(typeof openContextMenu, 'function');
+});